First pass at global and node time dilation (#142)

+++ editor/src/core/editor_context.cpp
@@ -20,5 +20,5 @@ std::string EditorContext::GetProjectTempPath() const {
 }
 
 float EditorContext::Time() {
-    return SDL_GetTicks() / 1000.0f;
+    return (float) SDL_GetTicks() / 1000.0f;
 }
